wildfire-risk-mvp/requirements.txt
# Core GEE
earthengine-api>=0.1.370

# Data & config
pyyaml>=6.0
numpy>=1.26.0

# Visualization
folium>=0.16.0

# API (optional)
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.29.0
